develop反映

+++ db/migrate/20201019004257_create_likes.rb
@@ -1,8 +1,8 @@
 class CreateLikes < ActiveRecord::Migration[5.2]
   def change
     create_table :likes do |t|
-      t.integer "user_id", null: false
-      t.integer "review_id", null: false
+      t.bigint "user_id", null: false
+      t.bigint "review_id", null: false
       t.timestamps
     end
   end
